+++ src/components/ConsortiumManagerHeader/ConsortiumManagerHeader.js
@@ -14,6 +14,7 @@ import {
 import { EVENT_EMITTER_EVENTS } from '../../constants';
 import { useConsortiumManagerContext } from '../../contexts';
 import { useEventEmitter } from '../../hooks';
+import { ConsortiumManagerNavigationPaneToggle } from '../ConsortiumManagerNavigationPaneToggle';
 import { FindConsortiumMember } from '../FindConsortiumMember';
 
 import css from './ConsortiumManagerHeader.css';
@@ -24,6 +25,7 @@ export const ConsortiumManagerHeader = () => {
   const [selectMembersDisabled, setSelectMembersDisabled] = useState();
   const {
     affiliations,
+    isNavigationPaneVisible,
     selectedMembers,
     selectMembers,
   } = useConsortiumManagerContext();
@@ -43,6 +45,12 @@ export const ConsortiumManagerHeader = () => {
     affiliations.map(({ tenantId, tenantName }) => ({ id: tenantId, name: tenantName }))
   ), [affiliations]);
 
+  const firstMenu = useMemo(() => {
+    if (selectMembersDisabled || isNavigationPaneVisible) return null;
+
+    return <ConsortiumManagerNavigationPaneToggle />;
+  }, [isNavigationPaneVisible, selectMembersDisabled]);
+
   return (
     <Paneset nested>
       <PaneHeader
@@ -56,6 +64,7 @@ export const ConsortiumManagerHeader = () => {
             values={{ amount: selectedMembers.length }}
           />
         )}
+        firstMenu={firstMenu}
         lastMenu={(
           <FindConsortiumMember
             disabled={selectMembersDisabled}