+++ src/components/index.js
@@ -1,5 +1,6 @@
 export { AffiliationLookupSuppressor } from './AffiliationLookupSuppressor';
 export { ConsortiaControlledVocabulary } from './ConsortiaControlledVocabulary';
 export { ConsortiumManagerHeader } from './ConsortiumManagerHeader';
+export { ConsortiumManagerNavigationPaneToggle } from './ConsortiumManagerNavigationPaneToggle';
 export { FindConsortiumMember } from './FindConsortiumMember';
 export { SwitchActiveAffiliation } from './SwitchActiveAffiliation';