+++ src/contexts/ConsortiumManagerContext.js
@@ -4,11 +4,10 @@ import {
   useCallback,
   useContext,
   useMemo,
+  useState,
 } from 'react';
 
-import {
-  useStripes,
-} from '@folio/stripes/core';
+import { useStripes } from '@folio/stripes/core';
 
 import {
   useCurrentUserTenantsPermissions,
@@ -23,6 +22,8 @@ export const ConsortiumManagerContext = createContext();
 export const ConsortiumManagerContextProvider = ({ children }) => {
   const stripes = useStripes();
   const userId = stripes?.user?.user?.id;
+
+  const [isNavigationPaneVisible, setIsNavigationPaneVisible] = useState(true);
 
   const {
     members,
@@ -69,14 +70,17 @@ export const ConsortiumManagerContextProvider = ({ children }) => {
     affiliations,
     hasPerm,
     isFetching,
+    isNavigationPaneVisible,
     permissionNamesMap,
     selectedMembers: members || DEFAULT_SELECTED_MEMBERS,
     selectMembers: updateMembersSelection,
+    setIsNavigationPaneVisible,
   }), [
     members,
     affiliations,
     hasPerm,
     isFetching,
+    isNavigationPaneVisible,
     permissionNamesMap,
     updateMembersSelection,
   ]);