+++ src/hooks/useUsers/index.js
@@ -1,1 +1,1 @@
-export { useUsers } from "./useUsers";
+export { useUsers } from './useUsers';