+++ src/routes/ConsortiumManager/settings/authorization-roles/UsersCapabilities/UsersCapabilitiesCompare/UsersCapabilitiesCompare.test.js
@@ -2,7 +2,7 @@ import { MemoryRouter } from 'react-router-dom';
 
 import {
   render,
-  screen
+  screen,
 } from '@folio/jest-config-stripes/testing-library/react';
 import { Paneset } from '@folio/stripes/components';