+++ src/routes/ConsortiumManager/settings/authorization-roles/UsersCapabilities/index.js
@@ -1,2 +1,2 @@
-export { UsersCapabilitiesCompare } from './UsersCapabilitiesCompare'
-export { UsersCapabilitiesCompareItems } from './UsersCapabilitiesCompareItems'
+export { UsersCapabilitiesCompare } from './UsersCapabilitiesCompare';
+export { UsersCapabilitiesCompareItems } from './UsersCapabilitiesCompareItems';